% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedRCBD-package.R
\docType{package}
\name{augmentedRCBD-package}
\alias{augmentedRCBD-package}
\title{augmentedRCBD: Analysis of Augmented Randomised Complete Block Designs}
\description{
Functions for analysis of data generated from experiments in augmented randomised complete block design according to Federer, W.T. (1961) \doi{10.2307/2527837}. Computes analysis of variance, adjusted means, descriptive statistics, genetic variability statistics etc. Further includes data visualization and report generation functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aravind-j/augmentedRCBD} \url{https://CRAN.R-project.org/package=augmentedRCBD} \url{https://aravind-j.github.io/augmentedRCBD/} \doi{10.5281/zenodo.1310011}
  \item Report bugs at \url{https://github.com/aravind-j/augmentedRCBD/issues}
}

}
\author{
\strong{Maintainer}: J. Aravind \email{j.aravind@icar.org.in} (\href{https://orcid.org/0000-0002-4791-442X}{ORCID})

Authors:
\itemize{
  \item J. Aravind \email{j.aravind@icar.org.in} (\href{https://orcid.org/0000-0002-4791-442X}{ORCID})
  \item S. Mukesh Sankar \email{mukesh.s@icar.org.in} (\href{https://orcid.org/0000-0001-5459-392X}{ORCID})
  \item Dhammaprakash Pandhari Wankhede \email{d.wankhede@icar.org.in} (\href{https://orcid.org/0000-0001-6384-8664}{ORCID})
  \item Vikender Kaur \email{vikender.kaur@icar.org.in} (\href{https://orcid.org/0000-0001-9698-7329}{ORCID})
}

Other contributors:
\itemize{
  \item ICAR-NBGPR (\href{https://ror.org/00scbd467}{ROR}) (url: https://nbpgr.org.in/) [copyright holder]
}

}
\keyword{internal}
