% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedRCBD.bulk.R
\name{augmentedRCBD.bulk}
\alias{augmentedRCBD.bulk}
\title{Analysis of Augmented Randomised Complete Block Design for Multiple
Traits/Characters}
\usage{
augmentedRCBD.bulk(data, block, treatment, traits, checks = NULL,
  method.comp = c("lsd", "tukey"), alpha = 0.05, describe = TRUE,
  freqdist = TRUE, gva = TRUE, round.digits)
}
\arguments{
\item{data}{}

\item{round.digits}{}
}
\description{
Analysis of Augmented Randomised Complete Block Design for Multiple
Traits/Characters
}
