% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedRCBD.menv.R
\name{augmentedRCBD.menv}
\alias{augmentedRCBD.menv}
\title{Combined Analysis of Augmented Randomised Complete Block Design in Multiple
Environments}
\usage{
augmentedRCBD.menv(
  block,
  treatment,
  env,
  y,
  checks = NULL,
  method.comp = c("lsd", "tukey", "none"),
  alpha = 0.05,
  group = TRUE,
  console = TRUE,
  simplify = FALSE,
  truncate.means = TRUE
)
}
\arguments{
\item{block}{Vector of blocks (as a factor).}

\item{treatment}{Vector of treatments/genotypes (as a factor).}

\item{env}{Vector of environments (as a factor).}

\item{y}{Numeric vector of response variable (Trait).}

\item{checks}{Character vector of the checks present in \code{treatment}
levels. If not specified, checks are inferred from the data on the basis of
number of replications of treatments/genotypes.}

\item{method.comp}{Method for comparison of treatments (\code{"lsd"} for
least significant difference or \code{"tukey"} for Tukey's honest
significant difference). If \code{"none"}, no comparisons will be made, the
ANOVA output will be given as a data frame and the adjusted means will be
computed directly from treatment and block effects instead of using
\code{\link[emmeans]{emmeans}}.}

\item{alpha}{Type I error probability (Significance level) to be used for
multiple comparisons.}

\item{group}{If \code{TRUE}, genotypes will be grouped according to
\code{"method.comp"}. Default is \code{TRUE}.}

\item{console}{If \code{TRUE}, output will be printed to console. Default is
\code{TRUE}. Default is \code{TRUE}.}

\item{simplify}{If \code{TRUE}, ANOVA output will be given as a data frame
instead of a \code{summary.aov} object. Default is \code{TRUE}.}

\item{truncate.means}{If \code{TRUE}, the negative adjusted means will be
truncated to zero. Default is \code{TRUE}.}
}
\value{
A list of class \code{augmentedRCBD.menv} containing the following
  components:  \item{\code{Details}}{Details of the augmented design used.}
 \item{\code{Means}}{A data frame with the "Means", "Block", "SE", "Mix",
 "Max" and "Adjusted Means" for each "Treatment".} \item{\code{ANOVA,
 Treatment Adjusted}}{An object of class \code{summary.aov} for ANOVA table
 with treatments adjusted.} \item{\code{ANOVA, Block Adjusted}}{An object of
 class \code{summary.aov} for ANOVA table with block adjusted.}
  \item{\code{Block effects}}{A vector of block effects.}
  \item{\code{Treatment effects}}{A vector of treatment effects.}
 \item{\code{Std. Errors}}{A data frame of standard error of difference
 between various combinations along with critical difference and tukey's
 honest significant difference (when \code{method.comp = "tukey"}) at
 \code{alpha}.} \item{\code{Overall adjusted mean}}{Overall adjusted mean.}
 \item{\code{CV}}{Coefficient of variation.} \item{\code{Comparison
 method}}{The method for comparison of treatments.}
 \item{\code{Comparisons}}{A data frame of pairwise comparisons of
 treatments. This is computed only if argument \code{group} is \code{TRUE}}
 \item{\code{Groups}}{A data frame with compact letter display of pairwise
 comparisons of treatments. Means with at least one letter common are not
 significantly different statistically. This is computed only if argument
 \code{group} is \code{TRUE} } \item{\code{warning}}{A vector of warning
 messages (if any) captured during model fitting. }
}
\description{
\code{augmentedRCBD.menv} is an extension of \code{augmentedRCBD} for the
combined/pooled analysis of data from augmented randomised complete block
design across multiple environments (locations and/or seasons).
}
