% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics-augmented}
\alias{generics-augmented}
\alias{describe}
\alias{freqdist}
\alias{gva}
\alias{report}
\title{Generic functions for \code{augmentedRCBD} family}
\usage{
describe(aug, ...)

freqdist(aug, ...)

gva(aug, ...)

report(aug, ...)
}
\arguments{
\item{aug}{An object of class \code{augmentedRCBD}, \code{augmentedRCBD.menv}
or \code{augmentedRCBD.mix}.}

\item{...}{Unused.}
}
\description{
Generic functions for \code{augmentedRCBD} family
}
\keyword{internal}
