% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gva.augmentedRCBD.menv.R
\name{gva.augmentedRCBD.menv}
\alias{gva.augmentedRCBD.menv}
\title{Perform Genetic Variability Analysis on \code{augmentedRCBD.menv} Output}
\usage{
\method{gva}{augmentedRCBD.menv}(aug, k = 2.063, ...)
}
\arguments{
\item{aug}{An object of class \code{augmentedRCBD.menv}.}

\item{k}{The standardized selection differential or selection intensity.
Default is 2.063 for 5\% selection proportion.}

\item{...}{Unused}
}
\value{
A list with the following descriptive statistics: \item{Mean}{The
  mean value (averaged across environments).} \item{PV}{Phenotyic variance.}
  \item{GV}{Genotyipc variance.} \item{EV}{Environmental variance.}
  \item{GCV}{Genotypic coefficient of variation} \item{GCV category}{The
  \mjseqn{GCV} category.} \item{PCV}{Phenotypic coefficient of variation}
  \item{PCV category}{The \mjseqn{PCV} category.} \item{ECV}{Environmental
  coefficient of variation} \item{hBS}{The broad-sense heritability
  (\mjseqn{H^{2}}).} \item{hBS category}{The \mjseqn{H^{2}} category.}
  \item{GA}{Genetic advance.} \item{GAM}{Genetic advance as per cent of mean.}
  \item{GAM category}{The \mjseqn{GAM} category.}
}
\description{
\code{gva.augmentedRCBD.menv} performs genetic variability analysis on an
object of class \code{augmentedRCBD.menv}. \loadmathjax
}
\details{
\code{gva.augmentedRCBD.menv} performs genetic variability analysis from the
ANOVA results in an object of class \code{augmentedRCBD.menv} and computes
several variability estimates. The analysis is based on the mean squares
averaged across environments as reported in the combined ANOVA table.

The phenotypic, genotypic and environmental variance
(\mjseqn{\sigma^{2}_{p}}, \mjseqn{\sigma^{2}_{g}} and \mjseqn{\sigma^{2}_{e}}
) are obtained from the ANOVA tables according to the expected value of mean
square described by Federer and Searle (1976) as follows:

\mjsdeqn{\sigma^{2}_{g} = \sigma^{2}_{p} - \sigma^{2}_{e}}

Phenotypic and genotypic coefficients of variation (\mjseqn{PCV} and
\mjseqn{GCV}) are estimated according to Burton (1951, 1952) as follows:

\mjsdeqn{GCV = \frac{\sigma^{2}_{g}}{\sqrt{\overline{x}}} \times 100}

Where \mjseqn{\overline{x}} is the mean (averaged across environments).

The estimates of \mjseqn{PCV} and \mjseqn{GCV} are categorised according to
Sivasubramanian and Madhavamenon (1978) as follows:

\tabular{ll}{ \strong{\emph{CV} (\%)} \tab \strong{Category} \cr x \mjteqn{ <
}{ &lt; }{ < } 10 \tab Low \cr 10 \mjseqn{\le} x \mjteqn{ < }{ &lt; }{ < } 20
\tab Medium \cr \mjseqn{\ge} 20 \tab High }

The broad-sense heritability (\mjseqn{H^{2}}) is calculated according to
method of Lush (1940) as follows:

\mjsdeqn{H^{2} = \frac{\sigma^{2}_{g}}{\sigma^{2}_{p}}}

The estimates of broad-sense heritability (\mjseqn{H^{2}}) are categorised
according to Robinson (1966) as follows:

\tabular{ll}{ \strong{\mjseqn{H^{2}}} \tab \strong{Category} \cr x \mjteqn{ <
}{ &lt; }{ < } 30 \tab Low \cr 30 \mjseqn{\le} x \mjteqn{ < }{ &lt; }{ < } 60
\tab Medium \cr \mjseqn{\ge} 60 \tab High }

Genetic advance (\mjseqn{GA}) is estimated and categorised according to
Johnson et al., (1955) as follows:

\mjsdeqn{GA = k \times \sigma_{g} \times \frac{H^{2}}{100}}

Where the constant \mjseqn{k} is the standardized selection differential or
selection intensity. The value of \mjseqn{k} at 5\% proportion selected is
2.063. Values of \mjseqn{k} at other selected proportions are available in
Appendix Table A of Falconer and Mackay (1996).

Genetic advance as per cent of mean (\mjseqn{GAM}) are estimated and
categorised according to Johnson et al., (1955) as follows:

\mjsdeqn{GAM = \frac{GA}{\overline{x}} \times 100}

\tabular{ll}{ \emph{\strong{GAM}} \tab \strong{Category} \cr x \mjseqn{ < }
10 \tab Low \cr 10 \mjseqn{\le} x \mjseqn{ < } 20 \tab Medium \cr
\mjseqn{\ge} 20 \tab High }
}
\note{
Genetic variability analysis needs to be performed only if the sum of
  squares of "Treatment: Test" are significant.

  Negative estimates of variance components if computed are not abnormal. For
  information on how to deal with these, refer Robinson (1955) and Dudley and
  Moll (1969).
}
\examples{

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Scenario 1: Test treatments are replicated across all environments
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Example data
blk1 <- c(1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3,
          4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 6, 6, 6,
          7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 8, 8, 8, 9, 9, 9, 9, 9, 9, 9)
trt1 <- c(1, 2, 3, 4, 7, 11, 12, 1, 2, 3, 4, 5, 9, 1, 2, 3, 4, 8, 6, 10,
          1, 2, 3, 4, 8, 11, 5, 1, 2, 3, 4, 12, 9, 1, 2, 3, 4, 7, 6, 10,
          1, 2, 3, 4, 7, 9, 12, 1, 2, 3, 4, 5, 6, 1, 2, 3, 4, 8, 11, 10)
y1 <- c(92, 79, 87, 81, 96, 89, 82, 79, 81, 81, 91, 79, 78, 83, 77,
        78, 78, 70, 75, 74, 90, 80, 85, 78, 95, 86, 81, 78, 78, 76, 88,
        76, 79, 80, 76, 75, 74, 77, 75, 72, 91, 81, 86, 80, 94, 87, 83,
        78, 79, 77, 90, 74, 76, 82, 83, 86, 76, 73, 74, 69)
env1 <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
          1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
          2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3,
          3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3)
data1 <- data.frame(env1, blk1, trt1, y1)
chks1 <- c(1, 2, 3, 4)

# Convert block, treatment and environment to factors
data1$blk1 <- as.factor(data1$blk1)
data1$trt1 <- as.factor(data1$trt1)
data1$env1 <- as.factor(data1$env1)

# Contingency tables of factors
table(data1$env1, data1$trt1)
table(data1$env1, data1$blk1)
table(data1$blk1, data1$trt1)

# Results
out1 <- augmentedRCBD.menv(block = data1$blk1, treatment = data1$trt1,
                           env = data1$env1, y = data1$y1, checks = chks1,
                           scenario = 1, method.comp = "lsd", alpha = 0.05,
                           group = TRUE, console = TRUE)

# Descriptive statistics
gva(out1)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Test treatments are not replicated across all environments
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Example data
blk2 <- c(1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3,
          4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 6, 6, 6,
          7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 8, 8, 8, 9, 9, 9, 9, 9, 9, 9)
trt2 <- c(1, 2, 3, 4, 7, 10, 11, 1, 2, 3, 4, 5, 9, 1, 2, 3, 4, 8, 6, 12,
          1, 2, 3, 4, 16, 19, 13, 1, 2, 3, 4, 20, 17, 1, 2, 3, 4, 15, 14, 18,
          1, 2, 3, 4, 22, 25, 27, 1, 2, 3, 4, 21, 23, 1, 2, 3, 4, 24, 26, 28)
y2 <- c(92, 79, 87, 81, 96, 89, 82, 79, 81, 81, 91, 79, 78, 83, 77,
        78, 78, 70, 75, 74, 90, 80, 85, 78, 95, 86, 81, 78, 78, 76, 88,
        76, 79, 80, 76, 75, 74, 77, 75, 72, 91, 81, 86, 80, 94, 87, 83,
        78, 79, 77, 90, 74, 76, 82, 83, 86, 76, 73, 74, 69)
env2 <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
          1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
          2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3,
          3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3)
data2 <- data.frame(env2, blk2, trt2, y2)
chks2 <- c(1, 2, 3, 4)

# Convert block, treatment and environment to factors
data2$blk2 <- as.factor(data2$blk2)
data2$trt2 <- as.factor(data2$trt2)
data2$env2 <- as.factor(data2$env2)

# Contingency tables of factors
table(data2$env2, data2$trt2)
table(data2$env2, data2$blk2)
table(data2$blk2, data2$trt2)

# Results
out2 <- augmentedRCBD.menv(block = data2$blk2, treatment = data2$trt2,
                           env = data2$env2, y = data2$y2, checks = chks2,
                           scenario = 2, method.comp = "lsd", alpha = 0.05,
                           group = TRUE, console = TRUE)

# Descriptive statistics
gva(out2)

}
\references{
\insertRef{lush_intra-sire_1940}{augmentedRCBD}

\insertRef{burton_quantitative_1951}{augmentedRCBD}

\insertRef{burton_qualitative_1952}{augmentedRCBD}

\insertRef{johnson_estimates_1955}{augmentedRCBD}

\insertRef{robinson_quantitative_1966}{augmentedRCBD}

\insertRef{federerModelConsiderationsVariance1976}{augmentedRCBD}

\insertRef{falconer_introduction_1996}{augmentedRCBD}
}
\seealso{
\code{\link[augmentedRCBD]{augmentedRCBD.menv}}
}
