% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.augmentedRCBD.bulk.R
\name{print.augmentedRCBD.bulk}
\alias{print.augmentedRCBD.bulk}
\title{Prints summary of \code{augmentedRCBD.bulk} object}
\usage{
\method{print}{augmentedRCBD.bulk}(x, ...)
}
\arguments{
\item{x}{An object of class \code{augmentedRCBD.bulk}.}

\item{...}{Unused}
}
\description{
\code{print.augmentedRCBD.bulk} prints to console the summary of an object of
class \code{augmentedRCBD.bulk} including the augmented design details,
trait-wise mean sum of squares from ANOVA (Treatment adjusted) and ANOVA
(Block adjusted), adjusted means, coefficient of variation, overall adjusted
means crtical differences, standard errors, descriptive statistics, frequency
distribution plots, genetic variability statistics and plots of genetic
variablity parameters.
}
\seealso{
\code{\link[augmentedRCBD]{augmentedRCBD.bulk}}
}
