% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.augmentedRCBD.menv.R
\name{print.augmentedRCBD.menv}
\alias{print.augmentedRCBD.menv}
\title{Prints summary of \code{augmentedRCBD.menv} object}
\usage{
\method{print}{augmentedRCBD.menv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{augmentedRCBD.menv}.}

\item{...}{Unused}
}
\description{
\code{print.augmentedRCBD.menv} prints to console the summary of an object of
class \code{augmentedRCBD.menv} including the augmented design details, ANOVA
(Treatment adjusted), ANOVA (Block adjusted), treatment means, coefficient of
variation, overall adjusted mean, critical differences and standard errors.
The treatment/genotype groups along with the grouping method are also printed
if they were computed.
}
\seealso{
\code{\link[augmentedRCBD]{augmentedRCBD.menv}}
}
