% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.augmentedRCBD.mix.R
\name{print.augmentedRCBD.mix}
\alias{print.augmentedRCBD.mix}
\title{Prints summary of \code{augmentedRCBD.mix} object}
\usage{
\method{print}{augmentedRCBD.mix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{augmentedRCBD.mix}.}

\item{...}{Unused}
}
\description{
\code{print.augmentedRCBD.mix} prints to console the summary of an object of
class \code{augmentedRCBD.mix} including the augmented design details, ANOVA
(Treatment adjusted), ANOVA (Block adjusted), treatment means, coefficient of
variation, overall adjusted mean, critical differences and standard errors.
The treatment/genotype groups along with the grouping method are also printed
if they were computed.
}
\seealso{
\code{\link[augmentedRCBD]{augmentedRCBD.mix}}
}
