% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.augmentedRCBD.R
\name{print.augmentedRCBD}
\alias{print.augmentedRCBD}
\title{Prints summary of \code{augmentedRCBD} object}
\usage{
\method{print}{augmentedRCBD}(x, ...)
}
\arguments{
\item{x}{An object of class \code{augmentedRCBD}.}

\item{...}{Unused}
}
\description{
\code{print.augmentedRCBD} prints to console the summary of an object of
class \code{augmentedRCBD} including the augmented design details, ANOVA
(Treatment adjusted), ANOVA (Block adjusted), treatment means, coefficient of
variation, overall adjusted mean, critical differences and standard errors.
The treatment/genotype groups along with the grouping method are also printed
if they were computed.
}
\seealso{
\code{\link[augmentedRCBD]{augmentedRCBD}}
}
