% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedRCBD.menv.R
\encoding{UTF-8}
\name{augmentedRCBD.menv}
\alias{augmentedRCBD.menv}
\title{Combined Analysis of Augmented Randomised Complete Block Design in Multiple
Environments}
\usage{
augmentedRCBD.menv(
  block,
  treatment,
  env,
  y,
  checks = NULL,
  method.comp = c("lsd", "tukey", "none"),
  scenario = c(1, 2),
  alpha = 0.05,
  group = TRUE,
  console = TRUE,
  simplify = FALSE,
  truncate.means = TRUE
)
}
\arguments{
\item{block}{Vector of blocks (as a factor).}

\item{treatment}{Vector of treatments/genotypes (as a factor).}

\item{env}{Vector of environments (as a factor).}

\item{y}{Numeric vector of response variable (Trait).}

\item{checks}{Character vector of the checks present in \code{treatment}
levels. If not specified, checks are inferred from the data on the basis of
number of replications of treatments/genotypes.}

\item{method.comp}{Method for comparison of treatments (\code{"lsd"} for
least significant difference or \code{"tukey"} for Tukey's honest
significant difference). If \code{"none"}, no comparisons will be made, the
ANOVA output will be given as a data frame and the adjusted means will be
computed directly from treatment and block effects instead of using
\code{\link[emmeans]{emmeans}}.}

\item{scenario}{Either \code{1} or \code{2} (see \strong{Details}).}

\item{alpha}{Type I error probability (Significance level) to be used for
multiple comparisons.}

\item{group}{If \code{TRUE}, genotypes will be grouped according to
\code{"method.comp"}. Default is \code{TRUE}.}

\item{console}{If \code{TRUE}, output will be printed to console. Default is
\code{TRUE}. Default is \code{TRUE}.}

\item{simplify}{If \code{TRUE}, ANOVA output will be given as a data frame
instead of a \code{summary.aov} object. Default is \code{TRUE}.}

\item{truncate.means}{If \code{TRUE}, the negative adjusted means will be
truncated to zero. Default is \code{TRUE}.}
}
\value{
A list of class \code{augmentedRCBD.menv} containing the following
  components:  \item{\code{Details}}{Details of the augmented design used.}
 \item{\code{Means}}{A data frame with the "Means", "Block", "SE", "Mix",
 "Max" and "Adjusted Means" for each "Treatment".} \item{\code{ANOVA,
 Treatment Adjusted}}{An object of class \code{summary.aov} for ANOVA table
 with treatments adjusted.} \item{\code{ANOVA, Block Adjusted}}{An object of
 class \code{summary.aov} for ANOVA table with block adjusted.}
  \item{\code{Block effects}}{A vector of block effects.}
  \item{\code{Treatment effects}}{A vector of treatment effects.}
 \item{\code{Std. Errors}}{A data frame of standard error of difference
 between various combinations along with critical difference and tukey's
 honest significant difference (when \code{method.comp = "tukey"}) at
 \code{alpha}.} \item{\code{Overall adjusted mean}}{Overall adjusted mean.}
 \item{\code{CV}}{Coefficient of variation.} \item{\code{Comparison
 method}}{The method for comparison of treatments.}
 \item{\code{Comparisons}}{A data frame of pairwise comparisons of
 treatments. This is computed only if argument \code{group} is \code{TRUE}}
 \item{\code{Groups}}{A data frame with compact letter display of pairwise
 comparisons of treatments. Means with at least one letter common are not
 significantly different statistically. This is computed only if argument
 \code{group} is \code{TRUE} } \item{\code{warning}}{A vector of warning
 messages (if any) captured during model fitting. }
}
\description{
\code{augmentedRCBD.menv} is an extension of
\code{\link[augmentedRCBD]{augmentedRCBD}} designed for the combined/pooled
analysis of data from augmented randomised complete block design across
multiple environments such as locations and/or seasons.
This function performs analysis under the following two scenarios. \describe{
\item{\emph{Scenario 1}:}{Replicated test treatments across environments.}
\item{\emph{Scenario 2}:}{Non-replicated test treatments across
environments.} }
}
\details{
The method of analysis is determined by whether the test treatments are
replicated across different environments as indicated by the \code{scenario}
argument. The design is balanced in terms of the checks being replicated
across all the blocks and environments.

\describe{ \item{Scenario \code{1}:}{The test treatments are replicated
across all environments. Here the Treatment \\u00D7 Environment interaction
is fully estimable as it is balanced.}\item{Scenario \code{2}:}{The test
treatments are replicated across all environments. Here the Treatment
\\u00D7 Environment interaction is only partially estimable as it is
unbalanced..} }
}
\note{
\itemize{ \item Data should preferably be balanced i.e. all the check
  genotypes should be present in all the blocks. If not, a warning is issued.
  \item There should not be any missing values. \item The number of test
  genotypes can vary within a block. \item When a large number of treatments
  or genotypes are involved, the analysis becomes memory intensive because
  calculating adjusted means necessitates the creation of a reference grid
  matrix by \code{emmeans}, which can grow exponentially in size }
}
\examples{

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Scenario 1: Test treatments are replicated across all environments
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Example data
blk1 <- c(1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3,
          4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 6, 6, 6,
          7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 8, 8, 8, 9, 9, 9, 9, 9, 9, 9)
trt1 <- c(1, 2, 3, 4, 7, 11, 12, 1, 2, 3, 4, 5, 9, 1, 2, 3, 4, 8, 6, 10,
          1, 2, 3, 4, 8, 11, 5, 1, 2, 3, 4, 12, 9, 1, 2, 3, 4, 7, 6, 10,
          1, 2, 3, 4, 7, 9, 12, 1, 2, 3, 4, 5, 6, 1, 2, 3, 4, 8, 11, 10)
y1 <- c(92, 79, 87, 81, 96, 89, 82, 79, 81, 81, 91, 79, 78, 83, 77,
        78, 78, 70, 75, 74, 90, 80, 85, 78, 95, 86, 81, 78, 78, 76, 88,
        76, 79, 80, 76, 75, 74, 77, 75, 72, 91, 81, 86, 80, 94, 87, 83,
        78, 79, 77, 90, 74, 76, 82, 83, 86, 76, 73, 74, 69)
env1 <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
          1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
          2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3,
          3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3)
data1 <- data.frame(env1, blk1, trt1, y1)
chks1 <- c(1, 2, 3, 4)

# Convert block, treatment and environment to factors
data1$blk1 <- as.factor(data1$blk1)
data1$trt1 <- as.factor(data1$trt1)
data1$env1 <- as.factor(data1$env1)

# Contingency tables of factors
table(data1$env1, data1$trt1)
table(data1$env1, data1$blk1)
table(data1$blk1, data1$trt1)

# Results
out1 <- augmentedRCBD.menv(block = data1$blk1, treatment = data1$trt1,
                           env = data1$env1, y = data1$y1, checks = chks1,
                           scenario = 1, method.comp = "lsd", alpha = 0.05,
                           group = TRUE, console = TRUE)

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# Test treatments are not replicated across all environments
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Example data
blk2 <- c(1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3,
          4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 6, 6, 6, 6, 6,
          7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 8, 8, 8, 9, 9, 9, 9, 9, 9, 9)
trt2 <- c(1, 2, 3, 4, 7, 10, 11, 1, 2, 3, 4, 5, 9, 1, 2, 3, 4, 8, 6, 12,
          1, 2, 3, 4, 16, 19, 13, 1, 2, 3, 4, 20, 17, 1, 2, 3, 4, 15, 14, 18,
          1, 2, 3, 4, 22, 25, 27, 1, 2, 3, 4, 21, 23, 1, 2, 3, 4, 24, 26, 28)
y2 <- c(92, 79, 87, 81, 96, 89, 82, 79, 81, 81, 91, 79, 78, 83, 77,
        78, 78, 70, 75, 74, 90, 80, 85, 78, 95, 86, 81, 78, 78, 76, 88,
        76, 79, 80, 76, 75, 74, 77, 75, 72, 91, 81, 86, 80, 94, 87, 83,
        78, 79, 77, 90, 74, 76, 82, 83, 86, 76, 73, 74, 69)
env2 <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
          1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
          2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3,
          3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3)
data2 <- data.frame(env2, blk2, trt2, y2)
chks2 <- c(1, 2, 3, 4)

# Convert block, treatment and environment to factors
data2$blk2 <- as.factor(data2$blk2)
data2$trt2 <- as.factor(data2$trt2)
data2$env2 <- as.factor(data2$env2)

# Contingency tables of factors
table(data2$env2, data2$trt2)
table(data2$env2, data2$blk2)
table(data2$blk2, data2$trt2)

# Results
out2 <- augmentedRCBD.menv(block = data2$blk2, treatment = data2$trt2,
                           env = data2$env2, y = data2$y2, checks = chks2,
                           scenario = 2, method.comp = "lsd", alpha = 0.05,
                           group = TRUE, console = TRUE)

}
\seealso{
\code{\link[augmentedRCBD]{augmentedRCBD}}
}
