% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentedRCBD.mix.R
\name{augmentedRCBD.mix}
\alias{augmentedRCBD.mix}
\title{Title}
\usage{
augmentedRCBD.mix(
  block,
  treatment,
  env = NULL,
  y,
  checks = NULL,
  env.random = TRUE,
  check.random = TRUE,
  test.random = TRUE,
  drop.nonsig.interaction = TRUE,
  scenario = c("I", "II"),
  scenario.violation.threshold = 0.1,
  df_method = c("kenward-roger", "satterthwaite"),
  console = TRUE
)
}
\arguments{
\item{console}{}
}
\description{
Title
}
