% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.augmentedRCBD.R
\name{describe.augmentedRCBD}
\alias{describe.augmentedRCBD}
\title{Compute Descriptive Statistics from \code{augmentedRCBD} Output}
\usage{
describe.augmentedRCBD(aug)
}
\arguments{
\item{aug}{An object of class \code{augmentedRCBD}.}
}
\value{
A list with the following descriptive statistics:  \item{Count}{The
  number of treatments/genotypes.} \item{Mean}{The mean value.}
  \item{Std.Error}{The standard error.} \item{Std.Deviation}{The standard
  deviation.} \item{Min}{The minimum value} \item{Max}{The maximum value}
  \item{Skewness(statistic)}{The skewness estimator.}
  \item{Skewness(p.value)}{The p-value from D'Agostino test of skewness.}
  \item{Kurtosis(statistic)}{The kurtosis estimator.}
  \item{Kurtosis(p.value)}{The p-value from Anscombe-Glynn test of kurtosis.}
}
\description{
\code{describe.augmentedRCBD} computes descriptive statistics from the
adjusted means in an object of class \code{augmentedRCBD}.
}
\details{
\code{describe.augmentedRCBD} computes the following descriptive statistics
from the adjusted means in an object of class \code{augmentedRCBD}.

\itemize{ \item Count \item Mean \item Standard deviation \item Standard
error \item Minimum \item Maximum \item Skewness statistic along with p-value
from D'Agostino test of skewness (D'Agostino, 1970). \item Kurtosis statistic
along with p-value from Anscombe-Glynn test of kurtosis (Anscombe and Glynn,
1983). }
}
\references{
\insertRef{dagostino_transformation_1970}{augmentedRCBD}

\insertRef{anscombe_distribution_1983}{augmentedRCBD}
}
\seealso{
\code{\link[augmentedRCBD]{augmentedRCBD}}
}
